Moved Media to customer package
