+++ src/main/java/com/project/trackfit/Main.java
@@ -1,11 +1,7 @@
 package com.project.trackfit;
 
-import com.project.trackfit.aws.S3Buckets;
-import com.project.trackfit.aws.S3Service;
-import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
 
 
 @SpringBootApplication
@@ -15,23 +11,4 @@ public class Main {
         SpringApplication.run(Main.class, args);
     }
 
-/*    @Bean
-    CommandLineRunner runner(S3Service s3Service, S3Buckets s3Buckets) {
-        return args -> testBucket(s3Service, s3Buckets);
-    }
-
-    private static void testBucket(S3Service s3Service, S3Buckets s3Buckets) {
-        s3Service.putObject(
-                s3Buckets.getCustomer(),
-                "foo",
-                "Hello World".getBytes()
-        );
-
-        byte[] object = s3Service.getObject(
-                s3Buckets.getCustomer(),
-                "foo"
-        );
-
-        System.out.println("Hooray " + new String(object));
-    }*/
 }
