+++ src/main/java/com/project/trackfit/customer/entity/Customer.java
@@ -3,7 +3,6 @@ package com.project.trackfit.customer.entity;
 import com.fasterxml.jackson.annotation.JsonManagedReference;
 import com.project.trackfit.user.entity.ApplicationUser;
 import com.project.trackfit.measurements.Measurements;
-import com.project.trackfit.media.Media;
 import com.project.trackfit.steps.DailySteps;
 import com.project.trackfit.subscription.Subscription;
 import jakarta.persistence.Column;
