+++ src/main/java/com/project/trackfit/customer/entity/Media.java
@@ -1,7 +1,6 @@
-package com.project.trackfit.media;
+package com.project.trackfit.customer.entity;
 
 import com.fasterxml.jackson.annotation.JsonBackReference;
-import com.project.trackfit.customer.entity.Customer;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
