+++ src/main/java/com/project/trackfit/customer/repository/MediaRepository.java
@@ -1,5 +1,6 @@
-package com.project.trackfit.media;
+package com.project.trackfit.customer.repository;
 
+import com.project.trackfit.customer.entity.Media;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
